requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langsmith>=0.6.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
openai>=1.3.0
anthropic>=0.8.0
PyPDF2>=3.0.1
python-docx>=1.1.0
openpyxl>=3.1.2
Pillow>=10.0.0
pytesseract>=0.3.10
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
httpx>=0.25.0